feat: add share dialog for video series page and podcast episode resource card (#3492)
* feat: Add Share/Embed dialog for Video Detail Page which should have video embed url and html


---------
