+++ frontends/main/src/app-pages/VideoPlaylistCollectionPage/VideoSeriesDetailPage.styled.ts
@@ -22,11 +22,45 @@ export const BreadcrumbBar = styled.div(({ theme }) => ({
   },
 }))
 
+export const ShareRow = styled.div(({ theme }) => ({
+  display: "flex",
+  justifyContent: "flex-end",
+  width: "100%",
+  [theme.breakpoints.down("sm")]: {
+    marginBottom: "32px",
+  },
+}))
+export const UpNextRight = styled.div({
+  display: "flex",
+  alignItems: "center",
+  gap: "24px",
+})
+
+export const ShareButton = styled.button(({ theme }) => ({
+  display: "flex",
+  alignItems: "center",
+  gap: "8px",
+  borderRadius: "4px",
+  padding: "14px 12px",
+  height: "32px",
+  border: `1px solid ${theme.custom.colors.silverGrayLight}`,
+  background: `${theme.custom.colors.white}`,
+  cursor: "pointer",
+  ...theme.typography.body2,
+  color: theme.custom.colors.darkGray1,
+  fontWeight: theme.typography.fontWeightMedium,
+  "&:hover": {
+    color: theme.custom.colors.red,
+  },
+  [theme.breakpoints.down("sm")]: {
+    width: "100%",
+    justifyContent: "center",
+  },
+}))
+
 // ── Series navigation bar ──
 
-export const SeriesNavBar = styled.div(({ theme }) => ({
-  borderBottom: `1px solid ${theme.custom.colors.lightGray2}`,
-}))
+export const SeriesNavBar = styled.div({})
 
 export const SeriesNavTopRow = styled(VideoContainer)(({ theme }) => ({
   display: "flex",
@@ -94,6 +128,7 @@ export const SeriesNavBottomRow = styled(VideoContainer)(({ theme }) => ({
   alignItems: "center",
   justifyContent: "space-between",
   padding: "16px 0 16px !important",
+  borderBottom: `1px solid ${theme.custom.colors.lightGray2}`,
   [theme.breakpoints.down("lg")]: {
     padding: "16px 16px 16px !important",
     flexDirection: "column",
